task3 unix

+++ lab4/task3/src/string/string.h
@@ -14,7 +14,7 @@ int split_space(char*, char***);
 int char_to_str(int, char**);
 int add_str(char**, int, ...);
 int add_chr(char**, int, ...);
-long to_str(long, char**);
+int to_str(int, char**);
 int to_int(char*);
 int is_empty_str(char*);
